refactor: domain 으로 패키지 명 변경
