+++ src/main/java/subway/repository/LineRepository.java
@@ -1,7 +1,7 @@
 package subway.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import subway.model.Line;
+import subway.domain.Line;
 
 public interface LineRepository extends JpaRepository<Line, Long>{
 }
